Refactor the eye tracking logic into EyeTracking, so that EyeTrackingDemo is only a wrapper for command purposes
